fixed more Compiler warnings
Former-commit-id: ed1ff70ccf36a4123af203d2b205948ad373de2e

+++ external/CocosBuilderReader/CCBActionManager.m
@@ -297,7 +297,7 @@
     [self setFirstFrameForNode:node sequenceProperty:seqProp tweenDuration:tweenDuration];
     
     NSArray* keyframes = [seqProp keyframes];
-    int numKeyframes = keyframes.count;
+    NSUInteger numKeyframes = keyframes.count;
     
     if (numKeyframes > 1)
     {
